[android] Fix versioned permissions checks (#6348)
# Why

Resolves https://github.com/expo/expo/issues/6339

# How

`ReactNativeActivity` now implements versioned `PermissionAwareActivity` interface.
Also, I've move permissions related method form `BaseExperienceActivity` to `ReactNativeActivity`.

# Test plan

Unfortunately, `et add-sdk-version --platform android` failed on a different task. However, I've copied my changes into a separate script:

```
MAJOR_ABI_VERSION='37'
ABI_VERSION='abi37_0_0'
NEWLINE='\
'
# Update classes that implement DefaultHardwareBackBtnHandler or PermissionAwareActivity
BACK_BUTTON_HANDLER_CLASS='com.facebook.react.modules.core.DefaultHardwareBackBtnHandler'
PERMISSION_AWARE_ACTIVITY_CLASS='com.facebook.react.modules.core.PermissionAwareActivity'
PERMISSION_LISTENER_CLASS='com.facebook.react.modules.core.PermissionListener'
find expoview/src/main/java/host/exp/exponent -iname '*.java' -type f -print0 | xargs -0 sed -i '' -e "s/ADD_NEW_SDKS_HERE/BEGIN_SDK_$MAJOR_ABI_VERSION$NEWLINE\ \ \ \ $ABI_VERSION\.$BACK_BUTTON_HANDLER_CLASS,$NEWLINE\ \ \ \ $ABI_VERSION\.$PERMISSION_AWARE_ACTIVITY_CLASS,$NEWLINE\ \ \ \ \/\/ END_SDK_$MAJOR_ABI_VERSION$NEWLINE\ \ \ \ \/\/ ADD_NEW_SDKS_HERE/" \
# Add implementation of PermissionAwareActivity.requestPermissions
find expoview/src/main/java/host/exp/exponent -iname '*.java' -type f -print0 | xargs -0 sed -i '' "s/ADD_NEW_PERMISSION_AWARE_ACTIVITY_IMPLEMENTATION_HERE/BEGIN_SDK_$MAJOR_ABI_VERSION$NEWLINE\ \ \ \ @Override$NEWLINE\ \ \ \ public\ void\ requestPermissions(String\[\]\ strings\,\ int\ i\,\ $ABI_VERSION\.$PERMISSION_LISTENER_CLASS\ permissionListener)\ {$NEWLINE\ \ \ \ \ \ super\.requestPermissions(strings\,\ i\,\ permissionListener::onRequestPermissionsResult);$NEWLINE\ \ \ \ }$NEWLINE\ \ \ \ \/\/ END_SDK_$MAJOR_ABI_VERSION$NEWLINE\ \ \ \ \/\/ ADD_NEW_PERMISSION_AWARE_ACTIVITY_IMPLEMENTATION_HERE/"
```
And run it in android folder. This script generated correct file: 
```
// Copyright 2015-present 650 Industries. All rights reserved.

package host.exp.exponent.experience;

// Implement for each version.
public class MultipleVersionReactNativeActivity extends ReactNativeActivity implements
    // The 4-space indentation is used by android-build-aar.sh.
    // WHEN_DISTRIBUTING_REMOVE_FROM_HERE
    // WHEN_PREPARING_SHELL_REMOVE_FROM_HERE
    // BEGIN_SDK_33
    abi33_0_0.com.facebook.react.modules.core.DefaultHardwareBackBtnHandler,
    // END_SDK_33
    // BEGIN_SDK_34
    abi34_0_0.com.facebook.react.modules.core.DefaultHardwareBackBtnHandler,
    // END_SDK_34
    // BEGIN_SDK_35
    abi35_0_0.com.facebook.react.modules.core.DefaultHardwareBackBtnHandler,
    // END_SDK_35
    // BEGIN_SDK_36
    abi36_0_0.com.facebook.react.modules.core.DefaultHardwareBackBtnHandler,
    abi36_0_0.com.facebook.react.modules.core.PermissionAwareActivity,
    // END_SDK_36
    // BEGIN_SDK_37
    abi37_0_0.com.facebook.react.modules.core.DefaultHardwareBackBtnHandler,
    abi37_0_0.com.facebook.react.modules.core.PermissionAwareActivity,
    // END_SDK_37
    // ADD_NEW_SDKS_HERE
    // WHEN_PREPARING_SHELL_REMOVE_TO_HERE
    // WHEN_DISTRIBUTING_REMOVE_TO_HERE
    com.facebook.react.modules.core.DefaultHardwareBackBtnHandler {

    // BEGIN_SDK_36
    @Override
    public void requestPermissions(String[] strings, int i, abi36_0_0.com.facebook.react.modules.core.PermissionListener permissionListener) {
      super.requestPermissions(strings, i, permissionListener::onRequestPermissionsResult);
    }
    // END_SDK_36
    // BEGIN_SDK_37
    @Override
    public void requestPermissions(String[] strings, int i, abi37_0_0.com.facebook.react.modules.core.PermissionListener permissionListener) {
      super.requestPermissions(strings, i, permissionListener::onRequestPermissionsResult);
    }
    // END_SDK_37
    // ADD_NEW_PERMISSION_AWARE_ACTIVITY_IMPLEMENTATION_HERE

}
```
