+++ android/expoview/src/main/java/host/exp/exponent/experience/BaseExperienceActivity.java
@@ -3,30 +3,18 @@
 package host.exp.exponent.experience;
 
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.content.pm.PermissionInfo;
-import android.os.Build;
 import android.os.Bundle;
-import android.os.Process;
 import android.util.Log;
 import android.util.Pair;
 
 import com.facebook.drawee.backends.pipeline.Fresco;
-import com.facebook.react.modules.core.PermissionAwareActivity;
-import com.facebook.react.modules.core.PermissionListener;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
-
-import org.json.JSONException;
-
-import androidx.annotation.NonNull;
 
 import javax.inject.Inject;
 
 import de.greenrobot.event.EventBus;
-import host.exp.exponent.ABIVersion;
 import host.exp.exponent.Constants;
-import host.exp.exponent.ExponentManifest;
 import host.exp.exponent.RNObject;
 import host.exp.exponent.di.NativeModuleDepsProvider;
 import host.exp.exponent.gcm.GcmRegistrationIntentService;
@@ -35,14 +23,10 @@ import host.exp.exponent.kernel.ExponentError;
 import host.exp.exponent.kernel.ExponentErrorMessage;
 import host.exp.exponent.kernel.Kernel;
 import host.exp.exponent.kernel.KernelConstants;
-import host.exp.exponent.kernel.services.ExpoKernelServiceRegistry;
 import host.exp.exponent.utils.AsyncCondition;
-import host.exp.exponent.utils.ScopedPermissionsRequester;
 import host.exp.expoview.Exponent;
 
-import static host.exp.exponent.utils.ScopedPermissionsRequester.EXPONENT_PERMISSIONS_REQUEST;
-
-public abstract class BaseExperienceActivity extends MultipleVersionReactNativeActivity implements PermissionAwareActivity {
+public abstract class BaseExperienceActivity extends MultipleVersionReactNativeActivity {
   private static String TAG = BaseExperienceActivity.class.getSimpleName();
 
   private static abstract class ExperienceEvent {
@@ -79,11 +63,6 @@ public abstract class BaseExperienceActivity extends MultipleVersionReactNativeA
 
   @Inject
   Kernel mKernel;
-
-  @Inject
-  protected ExpoKernelServiceRegistry mKernelServiceRegistry;
-
-  private ScopedPermissionsRequester mScopedPermissionsRequester;
 
   private long mOnResumeTime;
 
@@ -303,82 +282,6 @@ public abstract class BaseExperienceActivity extends MultipleVersionReactNativeA
       // Probably we need to either remove GCM entirely from the clients, or update the
       // implementation to ensure we only try to register in the foreground (like we do with FCM).
       Log.e(TAG, "Failed to register for GCM notifications", e);
-    }
-  }
-
-  // for getting scoped permission
-  @Override
-  public int checkPermission(final String permission, final int pid, final int uid) {
-    int globalResult = super.checkPermission(permission, pid, uid);
-
-    // only these permissions, which show a dialog to the user should be scoped.
-    boolean isDangerousPermission;
-    try {
-      PermissionInfo permissionInfo = getPackageManager().getPermissionInfo(permission, PackageManager.GET_META_DATA);
-      isDangerousPermission = (permissionInfo.protectionLevel & PermissionInfo.PROTECTION_DANGEROUS) != 0;
-    } catch (PackageManager.NameNotFoundException e) {
-      return PackageManager.PERMISSION_DENIED;
-    }
-
-    if (Constants.isStandaloneApp() || !isDangerousPermission) {
-      return globalResult;
-    }
-
-    if (globalResult == PackageManager.PERMISSION_GRANTED &&
-        mKernelServiceRegistry.getPermissionsKernelService().hasGrantedPermissions(permission, mExperienceId)) {
-      return PackageManager.PERMISSION_GRANTED;
-    } else {
-      return PackageManager.PERMISSION_DENIED;
-    }
-  }
-
-  // for getting global permission
-  @Override
-  public int checkSelfPermission(String permission) {
-    return super.checkPermission(permission, Process.myPid(), Process.myUid());
-  }
-
-  @Override
-  public boolean shouldShowRequestPermissionRationale(@NonNull String permission) {
-    // in scoped application we don't have `don't ask again` button
-    if (!Constants.isStandaloneApp() && checkSelfPermission(permission) == PackageManager.PERMISSION_GRANTED) {
-      return true;
-    }
-    return super.shouldShowRequestPermissionRationale(permission);
-  }
-
-
-  @Override
-  public void onRequestPermissionsResult(final int requestCode, final String[] permissions, @NonNull final int[] grantResults) {
-    if (requestCode == EXPONENT_PERMISSIONS_REQUEST) {
-      // TODO: remove once SDK 35 is deprecated
-      String sdkVersion = "0.0.0";
-      try {
-        sdkVersion = mManifest.getString(ExponentManifest.MANIFEST_SDK_VERSION_KEY);
-      } catch (JSONException e) {
-        e.printStackTrace();
-      }
-      if (ABIVersion.toNumber(sdkVersion) < ABIVersion.toNumber("36.0.0")) {
-        Exponent.getInstance().onRequestPermissionsResult(requestCode, permissions, grantResults);
-      } else {
-        if (permissions.length > 0 && grantResults.length == permissions.length && mScopedPermissionsRequester != null) {
-          mScopedPermissionsRequester.onRequestPermissionsResult(permissions, grantResults);
-          mScopedPermissionsRequester = null;
-        }
-      }
-    } else {
-      super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-    }
-
-  }
-
-  @Override
-  public void requestPermissions(final String[] permissions, final int requestCode, final PermissionListener listener) {
-    if (requestCode == EXPONENT_PERMISSIONS_REQUEST) {
-      mScopedPermissionsRequester = new ScopedPermissionsRequester(mExperienceId);
-      mScopedPermissionsRequester.requestPermissions(this, mManifest.optString(ExponentManifest.MANIFEST_NAME_KEY), permissions, listener);
-    } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-      super.requestPermissions(permissions, requestCode);
     }
   }
 }
